module renamed to payshield_pressure_test to avoid collisions
