+++ payshield_pressure_test/pressureTest.py
@@ -38,10 +38,10 @@ import time
 from logging.handlers import RotatingFileHandler
 from pathlib import Path
 
-from payshield.connector import PayConnector
-from payshield.decoders import DECODERS
-from payshield.runner import run_test
-from payshield.updater import UpdateChecker
+from payshield_pressure_test.connector import PayConnector
+from payshield_pressure_test.decoders import DECODERS
+from payshield_pressure_test.runner import run_test
+from payshield_pressure_test.updater import UpdateChecker
 
 VERSION = "1.5.4"
